rename SynchronizedGraphQLClient -> TheGraphClient, GraphQLClient as internal helper class
